moved fh tests to TestAllForecasters class
