+++ sktime/forecasting/tests/test_all_forecasters.py
@@ -512,6 +512,37 @@ class TestAllForecasters(ForecasterFixtureGenerator, QuickTester):
         actual = y_pred.index
         np.testing.assert_array_equal(actual, expected)
 
+    def test__y_and_cutoff(estimator_instance, n_columns):
+        """Check cutoff and _y."""
+        # check _y and cutoff is None after construction
+        f = estimator_instance
+
+        y = _make_series(n_columns=n_columns)
+        y_train, y_test = temporal_train_test_split(y, train_size=0.75)
+
+        # check that _y and cutoff are empty when estimator is constructed
+        assert f._y is None
+        assert f.cutoff is None
+
+        # check that _y and cutoff is updated during fit
+        f.fit(y_train, fh=FH0)
+        # assert isinstance(f._y, pd.Series)
+        # action:uncomments the line above
+        # why: fails for multivariates cause they are DataFrames
+        # solution: look for a general solution for Series and DataFrames
+        assert len(f._y) > 0
+        assert f.cutoff == y_train.index[-1]
+
+        # check data pointers
+        np.testing.assert_array_equal(f._y.index, y_train.index)
+
+        # check that _y and cutoff is updated during update
+        f.update(y_test, update_params=False)
+        np.testing.assert_array_equal(
+            f._y.index, np.append(y_train.index, y_test.index)
+        )
+        assert f.cutoff == y_test.index[-1]
+
     def test__y_when_refitting(self, estimator_instance, n_columns):
         """Test that _y is updated when forecaster is refitted."""
         y_train = _make_series(n_columns=n_columns)
@@ -520,3 +551,48 @@ class TestAllForecasters(ForecasterFixtureGenerator, QuickTester):
         # using np.squeeze to make the test flexible to shape differeces like
         # (50,) and (50, 1)
         assert np.all(np.squeeze(estimator_instance._y) == np.squeeze(y_train[3:]))
+
+    def test_fh_attribute(self, estimator_instance, n_columns):
+        """Check fh attribute and error handling if two different fh are passed."""
+        f = estimator_instance
+        y_train = _make_series(n_columns=n_columns)
+
+        f.fit(y_train, fh=FH0)
+        np.testing.assert_array_equal(f.fh, FH0)
+        f.predict()
+        np.testing.assert_array_equal(f.fh, FH0)
+        f.predict(FH0)
+        np.testing.assert_array_equal(f.fh, FH0)
+
+        # if fh is not required in fit, test this again with fh passed late
+        if not f.get_tag("requires-fh-in-fit"):
+            f.fit(y_train)
+            f.predict(FH0)
+            np.testing.assert_array_equal(f.fh, FH0)
+
+    def test_fh_not_passed_error_handling(self, estimator_instance, n_columns):
+        """Check that not passing fh in fit/predict raises correct error."""
+        f = estimator_instance
+        y_train = _make_series(n_columns=n_columns)
+
+        if f.get_tag("requires-fh-in-fit"):
+            # if fh required in fit, should raise error if not passed in fit
+            with pytest.raises(ValueError):
+                f.fit(y_train)
+        else:
+            # if fh not required in fit, should raise error if not passed until predict
+            f.fit(y_train)
+            with pytest.raises(ValueError):
+                f.predict()
+
+    def test_different_fh_in_fit_and_predict_error_handling(
+        self, estimator_instance, n_columns
+    ):
+        """Check that fh different in fit and predict raises correct error."""
+        f = estimator_instance
+        y_train = _make_series(n_columns=n_columns)
+        f.fit(y_train, fh=FH0)
+        np.testing.assert_array_equal(f.fh, FH0)
+        # changing fh during predict should raise error
+        with pytest.raises(ValueError):
+            f.predict(fh=FH0 + 1)
